Avoid SourcePrinter to calculate newline while updating cursor Position
... by disallow passing strings with newlines to the sourceprinter.
